iceberg_datafusion::physical_plan

Function convert_filters_to_predicate

pub fn convert_filters_to_predicate(filters: &[Expr]) -> Option<Predicate>

Converts DataFusion filters ([Expr]) to an iceberg Predicate. If none of the filters could be converted, return None which adds no predicates to the scan operation. If the conversion was successful, return the converted predicates combined with an AND operator.